+++ index.html
@@ -234,7 +234,8 @@
       </section>
     </template>
 
-    <!-- <script src="js/functions.js"></script> -->
-    <script src="js/main.js" defer></script>
+    <script src="js/functions.js"></script>
+    <script src="js/main.js" type="module" defer></script>
+    <script src="js/data.js" type="module"></script>
 </body>
 </html>
