+++ js/setup.js
@@ -1,0 +1,43 @@
+const numberCount = 25;
+const commentCount = 30;
+const minLikesCount = 15;
+const maxLikesCount = 200;
+
+const names = [
+  'Никита',
+  'Егор',
+  'Виктория',
+  'Дарья',
+  'Арина',
+  'Григорий',
+  'Елизавета',
+  'Марина',
+  'Алексей',
+  'Алиса',
+  'Кира',
+  'Ева',
+  'Артём',
+  'Василиса',
+  'Анна'
+];
+
+const messageUser = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const description = [
+  'Красивый город',
+  'Фото на пляже',
+  'Вкусная еда',
+  'Море, солнце, пляж',
+  'Крутая машина',
+  'Котейка',
+  'Отдых на море'
+];
+
+export default [];
